Update PaymentRequest API for Chrome Dev 53.0.2774.4
The format of the Android Pay parameters has changed to pass in more
information or Android Pay. See the "Android Pay" sample for details.

PaymentRequest now allows merchants to collect user's contact
information: phone number and/or email address. The merchant can specify
this in options:

var options = {
  requestPayerPhone: true,
  requestPayerEmail: true
};

Chrome returns the contact information in the 'payerPhone' and
'payerEmail' fields in the instrument response.

The parameter type for the complete() method has changed from a boolean
to a string with two predefined values.

1) "success" - merchant successfully completed the transaction.
2) "fail" - merchant is not able to complete the transaction.

Any other string means the merchant cannot indicate either success or
failure to the user at the moment.

Chrome no longer returns the 'totalAmount' in the instrument response.